+++ game/src/camera.rs
@@ -1,0 +1,59 @@
+use common::{
+    glam::{Quat, Vec3},
+    Camera,
+};
+use components::Transform;
+
+use crate::{
+    input::{Input, Keys},
+    Game, CAMERA_ROTATE_SPEED, CAMERA_ZOOM_SPEED,
+};
+
+pub fn update_camera(game: &mut Game) {
+    let camera = &mut game.camera;
+    camera.target = game.world.get::<&Transform>(game.dave).unwrap().position;
+    let input = &game.input;
+    let dt = game.time.delta();
+
+    let focus_radius = 1.0;
+    let focus_centering = 0.5;
+    let distance_to_target = camera.target.distance(camera.focus_point);
+
+    let mut t = 1.0;
+    if distance_to_target > 0.01 {
+        t = ((1. - focus_centering) as f32).powf(dt);
+    }
+    if distance_to_target > focus_radius {
+        t = t.min(focus_radius / distance_to_target);
+    }
+    camera.focus_point = camera.target.lerp(camera.focus_point, t);
+
+    let camera_rotate = input.keyboard_state.as_axis(Keys::E, Keys::Q);
+    camera.yaw += camera_rotate * CAMERA_ROTATE_SPEED * dt;
+
+    set_camera_distance(input, camera, dt);
+
+    camera.pitch = -45_f32.to_radians();
+    let look_rotation = Quat::from_euler(common::glam::EulerRot::YXZ, camera.yaw, camera.pitch, 0.);
+    let look_direction = look_rotation * Vec3::NEG_Z;
+    let look_position = camera.focus_point - look_direction * camera.distance;
+
+    camera.position = look_position;
+}
+
+fn set_camera_distance(input: &Input, camera: &mut Camera, dt: f32) {
+    if input.camera_zoom.abs() > 0. {
+        camera.start_distance = camera.distance;
+        camera.desired_distance += input.camera_zoom;
+        camera.desired_distance = camera.desired_distance.clamp(5., 50.);
+    }
+
+    let current_delta = camera.desired_distance - camera.distance;
+
+    let epsilon = 0.01;
+    if current_delta.abs() > epsilon {
+        camera.distance += current_delta * CAMERA_ZOOM_SPEED * dt;
+    } else {
+        camera.distance = camera.desired_distance;
+    }
+}
