+++ game/src/input.rs
@@ -1,13 +1,78 @@
-use crate::ClickState;
-
-use super::{Game, Keys};
+use super::Game;
 use common::{
+    bitflags::bitflags,
+    glam::Vec2,
     log,
     winit::{
         self,
         event::{ElementState, KeyboardInput, VirtualKeyCode, WindowEvent},
     },
 };
+
+bitflags! {
+    #[derive(Default, Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+    pub struct Keys: u8 {
+        const W = 0b00000001;
+        const A = 0b00000010;
+        const S = 0b00000100;
+        const D = 0b00001000;
+        const Q = 0b00010000;
+        const E = 0b00100000;
+        const C = 0b01000000;
+        const Space = 0b10000000;
+    }
+}
+
+impl Keys {
+    pub fn as_axis(&self, negative: Keys, positive: Keys) -> f32 {
+        let negative = self.contains(negative) as i8 as f32;
+        let positive = self.contains(positive) as i8 as f32;
+        positive - negative
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct MouseState {
+    pub position: Option<Vec2>,
+    pub left_click_state: ClickState,
+    pub right_click_state: ClickState,
+    pub middle_click_state: ClickState,
+}
+
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub enum ClickState {
+    #[default]
+    Released,
+    Down,
+    JustReleased,
+}
+
+#[derive(Clone, Debug)]
+pub struct Input {
+    pub keyboard_state: Keys,
+    pub mouse_state: MouseState,
+    pub camera_zoom: f32,
+}
+
+impl Default for Input {
+    fn default() -> Self {
+        Self {
+            mouse_state: Default::default(),
+            keyboard_state: Default::default(),
+            camera_zoom: 0.,
+        }
+    }
+}
+
+impl Input {
+    pub fn reset(&mut self) {
+        *self = Default::default();
+    }
+
+    pub fn is_pressed(&self, key: Keys) -> bool {
+        self.keyboard_state.contains(key)
+    }
+}
 
 pub fn handle_winit_event(game: &mut Game, event: winit::event::WindowEvent) {
     match event {
@@ -28,6 +93,21 @@ pub fn handle_winit_event(game: &mut Game, event: winit::event::WindowEvent) {
         }
         _ => {}
     }
+}
+
+pub fn reset_mouse_clicks(mouse_state: &mut MouseState) {
+    match mouse_state.left_click_state {
+        ClickState::JustReleased => mouse_state.left_click_state = ClickState::Released,
+        _ => {}
+    };
+    match mouse_state.right_click_state {
+        ClickState::JustReleased => mouse_state.right_click_state = ClickState::Released,
+        _ => {}
+    };
+    match mouse_state.middle_click_state {
+        ClickState::JustReleased => mouse_state.middle_click_state = ClickState::Released,
+        _ => {}
+    };
 }
 
 fn handle_mousewheel(game: &mut Game, delta: winit::event::MouseScrollDelta) {
